Created views for login, register, logout
Created 3 views: "login_user", "register_user", "logout_user". Both views for login and register check if the request method is other than POST, and then an empty form is added to the context. If the method is GET, then data from the POST request is collected. The form is checked for validity. In the login view, the user is logged in if the form is valid, and in the register view, a new user is created and then logged in. Every view redirects to the home page after success.