Added logic for displaying private tasks
When the user is logged in, two queries are sent to the database: one for tasks that have passed their deadlines and another for tasks with deadlines set for today or in the future. If the user is not logged in, the data is retrieved from the session. This session data is then converted into QuerySets using an "anonymous" user. The QuerySets are divided into two lists: tasks_ended and tasks, following the same logic as when the user is logged in. In both cases, the tasks are grouped by date and stored in the "grouped_tasks" dictionary. Template "mytodo.hmtl" was modified to display dynamic data.

Small fixes:
- Added min-height to the CSS style for the "single-task" class to display long descriptions properly