Added  IDs for tasks stored in session
To simplify operations on tasks stored in session and to prevent potential issues with duplicated IDs in the future, a system for generating unique identifiers has been introduced. Instead of using simple incremental integers, UUIDs (Universally Unique Identifiers) are now employed as identifiers for tasks. This change ensures robust identification of tasks and mitigates the risk of conflicts caused by deleted records or other potential scenarios.