Added "delete_view"
If a user is logged in, the record will be deleted from the database. If the user is not logged in, the private tasks will be retrieved from the session, and every task will be checked until the task with the target ID is found. Then, the old data is replaced by the new data without the record to be deleted.