Added "complete_task"
If a user is logged in, the value of the "done" field will be changed to True for a specific record in the database. If the user is not logged in, the data will be taken from the session, similarly to how it works in "delete_task", and the data will be updated accordingly.

Small FIxes:
- Updated logic for displaying data from the session to show only tasks that are not marked as completed