Created view for displaying projects
Created a view for users to browse projects they are participating in. Added a template for displaying projects. Projects are now shown as tiles in a grid object and include example data. View is only available for users that are logged in. If user that is not logged tries to entered "todoproject" view he will be redirected to the login url.