Added form for new project in "todoproject.html"
Added 2 icons that will work as buttons to the template. First is notifcation icon that has number above it. That number will represent how many notifications user has. Second icon is for adding new project. When icon is clicked it will open form with input for project name. Form is now only in html and does not have real functionality.