Added notification tile to the "todoproject.html"
Added notifcation tile to the "todoproject.html". The tile is initially hidden and will only appear when the user clicks on the notification icon. The tile will display two kinds of messages: informative notifications, where the user cannot take any action, and question-type notifications, where the user can choose to either accept or decline. Tile does not display real notifications, but this functionality will be implemented in the future.