Added form and functionality for new projects
Integrated the form from "forms.py" into the 'todoproject' view. The form is based on the 'Project' model. Upon receiving a POST request, the form data is validated. If all data is correct, a new project is saved to the database. In case of any errors encountered during form validation, an error message is displayed to the user.