Mofied "todoproject" view to take real data
Updated "todoproject" view to dynamically generate data from the database for display on the page. The view retrieves all projects that the current user is a member of, along with their associated tables, and passes this data to the template as a list. Added to template one if that will display information if project does not have any table.