Added form to add tasks to tables
Created form with the hidden input to collect all data that is required to create new task in projects. The hidden input field is used to store the table ID, which is crucial for creating tasks within the correct project table. The value of the table ID is automatically populated by the JavaScript script, ensuring accuracy and consistency of data.