Added option to add new task
This commit introduces the ability to add new tasks to project tables. The form collects all required data for the new task, except for two pieces of information. The table ID is automatically filled by a JavaScript script, while the project ID is transferred via the URL. User currently can not choose what user is responsible for the task, but it will be changed in future.