Added AJAX funct for inviting users to projects.
Added functionality to the "add_user" view. View now process data and checks these things before sending invitation. If user exists in db, if user is not member already and if user was not invited already.  If any of these conditions are true, the view aborts the invitation process and returns appropriate feedback to the user, explaining the reason for rejection.